+++ packages/cli/src/process/runner.ts
@@ -4,7 +4,6 @@ type InitMessage = {
   init: true;
   basePath: string;
   opts: Opts;
-  //command?: string; // TODO execute | compile | validate etc
 };
 
 // When receiving a message as a child process, we pull out the args and run
@@ -15,3 +14,8 @@ process.on('message', ({ init, basePath, opts }: InitMessage) => {
     });
   }
 });
+
+// Tell the parent process we're awake and ready
+process.send!({
+  init: true
+})